Refactor slice elements
