+++ xtrack/beam_elements/slice_elements_drift.py
@@ -2,46 +2,19 @@ import xobjects as xo
 
 from ..general import _pkg_root
 from ..base_element import BeamElement
+from .slice_base import _SliceBase, COMMON_SLICE_XO_FIELDS
 from .elements import (
     SynchrotronRadiationRecord, Bend, Quadrupole, Sextupole,
     Octupole, Solenoid, Drift, RBend, UniformSolenoid
 )
-from ..random import RandomUniformAccurate, RandomExponential
 from ..survey import advance_element as survey_advance_element
 
-from .slice_elements_thin import _slice_copy, ID_RADIATION_FROM_PARENT
+class _DriftSliceElementBase(_SliceBase):
 
-
-COMMON_SLICE_XO_FIELDS = {
-    'radiation_flag': xo.Field(xo.Int64, default=ID_RADIATION_FROM_PARENT),
-    'delta_taper': xo.Float64,
-    'weight': xo.Float64,
-}
-
-class _DriftSliceElementBase:
-
-    allow_rot_and_shift = False
-    allow_loss_refinement = True
     rot_and_shift_from_parent = False
-    _skip_in_to_dict = ['_parent']
-    has_backtrack = True
-    _force_moveable = True
-    isthick = True
+    allow_loss_refinement = True
+    isthick=True
     _inherit_strengths = False
-
-    copy = _slice_copy
-
-    def to_dict(self, **kwargs):
-        dct = BeamElement.to_dict(self, **kwargs)
-        dct['parent_name'] = self.parent_name
-        return dct
-
-    @classmethod
-    def from_dict(cls, dct, **kwargs):
-        obj = super().from_dict(dct, **kwargs)
-        obj.parent_name = dct['parent_name']
-        return obj
-
 
 class DriftSliceBend(_DriftSliceElementBase, BeamElement):
 
@@ -55,7 +28,6 @@ class DriftSliceBend(_DriftSliceElementBase, BeamElement):
         out = Drift(length=self._parent.length * self.weight,
                      _buffer=self._buffer)
         return out
-
 
 class DriftSliceRBend(_DriftSliceElementBase, BeamElement):
 
@@ -95,7 +67,6 @@ class DriftSliceRBend(_DriftSliceElementBase, BeamElement):
 
         return v, w
 
-
 class DriftSliceQuadrupole(_DriftSliceElementBase, BeamElement):
 
     _xofields = {'_parent': xo.Ref(Quadrupole), **COMMON_SLICE_XO_FIELDS}
@@ -108,7 +79,6 @@ class DriftSliceQuadrupole(_DriftSliceElementBase, BeamElement):
         out = Drift(length=self._parent.length * self.weight,
                      _buffer=self._buffer)
         return out
-
 
 class DriftSliceSextupole(_DriftSliceElementBase, BeamElement):
 
@@ -123,7 +93,6 @@ class DriftSliceSextupole(_DriftSliceElementBase, BeamElement):
                      _buffer=self._buffer)
         return out
 
-
 class DriftSliceOctupole(_DriftSliceElementBase, BeamElement):
 
     _xofields = {'_parent': xo.Ref(Octupole), **COMMON_SLICE_XO_FIELDS}
@@ -136,7 +105,6 @@ class DriftSliceOctupole(_DriftSliceElementBase, BeamElement):
         out = Drift(length=self._parent.length * self.weight,
                      _buffer=self._buffer)
         return out
-
 
 class DriftSlice(_DriftSliceElementBase, BeamElement):
     _xofields = {'_parent': xo.Ref(Drift), **COMMON_SLICE_XO_FIELDS}
