+++ xtrack/beam_elements/slice_elements_thin.py
@@ -1,7 +1,7 @@
 import xobjects as xo
 
 from ..general import _pkg_root
-from ..random import RandomUniformAccurate, RandomExponential
+from .slice_base import _SliceBase, COMMON_SLICE_XO_FIELDS, ID_RADIATION_FROM_PARENT
 from .elements import (
     SynchrotronRadiationRecord, Quadrupole, Sextupole,
     Octupole, Bend, Multipole, DipoleEdge, RBend, MultipoleEdge, Marker,
@@ -9,49 +9,21 @@ from .elements import (
 )
 from ..base_element import BeamElement
 
-ID_RADIATION_FROM_PARENT = 10
+class _ThinSliceElementBase(_SliceBase):
 
-_common_xofields = {
-    'radiation_flag': xo.Field(xo.Int64, default=ID_RADIATION_FROM_PARENT),
-    'delta_taper': xo.Float64,
-    'weight': xo.Float64,
-}
-
-def _slice_copy(self, **kwargs):
-    out = BeamElement.copy(self, **kwargs)
-    out._parent = None
-    out.parent_name = self.parent_name
-    return out
-
-
-class ThinSliceQuadrupole(BeamElement):
-    allow_rot_and_shift = False
     rot_and_shift_from_parent = True
-    _skip_in_to_dict = ['_parent']
-    _depends_on = [RandomUniformAccurate, RandomExponential]
-    _internal_record_class = SynchrotronRadiationRecord
-    has_backtrack = True
-    _force_moveable = True
+    allow_loss_refinement = False
+    isthick = False
     _inherit_strengths = True
 
-    _xofields = {'_parent': xo.Ref(Quadrupole), **_common_xofields}
+
+class ThinSliceQuadrupole(_ThinSliceElementBase, BeamElement):
+
+    _xofields = {'_parent': xo.Ref(Quadrupole), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
         _pkg_root.joinpath('beam_elements/elements_src/thin_slice_quadrupole.h'),
     ]
-
-    copy = _slice_copy
-
-    def to_dict(self, **kwargs):
-        dct = BeamElement.to_dict(self, **kwargs)
-        dct['parent_name'] = self.parent_name
-        return dct
-
-    @classmethod
-    def from_dict(cls, dct, **kwargs):
-        obj = super().from_dict(dct, **kwargs)
-        obj.parent_name = dct['parent_name']
-        return obj
 
     def get_equivalent_element(self):
 
@@ -80,34 +52,13 @@ class ThinSliceQuadrupole(BeamElement):
         return out
 
 
-class ThinSliceSextupole(BeamElement):
-    allow_rot_and_shift = False
-    rot_and_shift_from_parent = True
-    _skip_in_to_dict = ['_parent']
-    _depends_on = [RandomUniformAccurate, RandomExponential]
-    _internal_record_class = SynchrotronRadiationRecord
-    has_backtrack = True
-    _force_moveable = True
-    _inherit_strengths = True
+class ThinSliceSextupole(_ThinSliceElementBase, BeamElement):
 
-    _xofields = {'_parent': xo.Ref(Sextupole), **_common_xofields}
+    _xofields = {'_parent': xo.Ref(Sextupole), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
         '#include <beam_elements/elements_src/thin_slice_sextupole.h>'
     ]
-
-    copy = _slice_copy
-
-    def to_dict(self, **kwargs):
-        dct = BeamElement.to_dict(self, **kwargs)
-        dct['parent_name'] = self.parent_name
-        return dct
-
-    @classmethod
-    def from_dict(cls, dct, **kwargs):
-        obj = super().from_dict(dct, **kwargs)
-        obj.parent_name = dct['parent_name']
-        return obj
 
     def get_equivalent_element(self):
 
@@ -136,34 +87,13 @@ class ThinSliceSextupole(BeamElement):
         return out
 
 
-class ThinSliceOctupole(BeamElement):
-    allow_rot_and_shift = False
-    rot_and_shift_from_parent = True
-    _skip_in_to_dict = ['_parent']
-    _depends_on = [RandomUniformAccurate, RandomExponential]
-    _internal_record_class = SynchrotronRadiationRecord
-    has_backtrack = True
-    _force_moveable = True
-    _inherit_strengths = True
+class ThinSliceOctupole(_ThinSliceElementBase, BeamElement):
 
-    _xofields = {'_parent': xo.Ref(Octupole), **_common_xofields}
+    _xofields = {'_parent': xo.Ref(Octupole), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
         '#include <beam_elements/elements_src/thin_slice_octupole.h>'
     ]
-
-    copy = _slice_copy
-
-    def to_dict(self, **kwargs):
-        dct = BeamElement.to_dict(self, **kwargs)
-        dct['parent_name'] = self.parent_name
-        return dct
-
-    @classmethod
-    def from_dict(cls, dct, **kwargs):
-        obj = super().from_dict(dct, **kwargs)
-        obj.parent_name = dct['parent_name']
-        return obj
 
     def get_equivalent_element(self):
 
@@ -192,34 +122,13 @@ class ThinSliceOctupole(BeamElement):
         return out
 
 
-class ThinSliceBend(BeamElement):
-    allow_rot_and_shift = False
-    rot_and_shift_from_parent = True
-    _skip_in_to_dict = ['_parent']
-    _depends_on = [RandomUniformAccurate, RandomExponential]
-    _internal_record_class = SynchrotronRadiationRecord
-    has_backtrack = True
-    _force_moveable = True
-    _inherit_strengths = True
+class ThinSliceBend(_ThinSliceElementBase, BeamElement):
 
-    _xofields = {'_parent': xo.Ref(Bend), **_common_xofields}
+    _xofields = {'_parent': xo.Ref(Bend), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
         '#include <beam_elements/elements_src/thin_slice_bend.h>'
     ]
-
-    copy = _slice_copy
-
-    def to_dict(self, **kwargs):
-        dct = BeamElement.to_dict(self, **kwargs)
-        dct['parent_name'] = self.parent_name
-        return dct
-
-    @classmethod
-    def from_dict(cls, dct, **kwargs):
-        obj = super().from_dict(dct, **kwargs)
-        obj.parent_name = dct['parent_name']
-        return obj
 
     def get_equivalent_element(self):
         knl = self._parent.knl.copy() * self.weight
@@ -246,36 +155,13 @@ class ThinSliceBend(BeamElement):
                         _buffer=self._buffer)
         return out
 
+class ThinSliceRBend(_ThinSliceElementBase, BeamElement):
 
-
-class ThinSliceRBend(BeamElement):
-    allow_rot_and_shift = False
-    rot_and_shift_from_parent = True
-    _skip_in_to_dict = ['_parent']
-    _depends_on = [RandomUniformAccurate, RandomExponential]
-    _internal_record_class = SynchrotronRadiationRecord
-    has_backtrack = True
-    _force_moveable = True
-    _inherit_strengths = True
-
-    _xofields = {'_parent': xo.Ref(RBend), **_common_xofields}
+    _xofields = {'_parent': xo.Ref(RBend), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
         '#include <beam_elements/elements_src/thin_slice_rbend.h>'
     ]
-
-    copy = _slice_copy
-
-    def to_dict(self, **kwargs):
-        dct = BeamElement.to_dict(self, **kwargs)
-        dct['parent_name'] = self.parent_name
-        return dct
-
-    @classmethod
-    def from_dict(cls, dct, **kwargs):
-        obj = super().from_dict(dct, **kwargs)
-        obj.parent_name = dct['parent_name']
-        return obj
 
     def get_equivalent_element(self):
         knl = self._parent.knl.copy() * self.weight
